+++ CHANGELOG.md
@@ -15,6 +15,7 @@ HEAD
 * Renamed `JsonObject::getOrCreate()` to `getOrAddMember()`
 * Fixed `JsonVariant::isNull()` not returning `true` after `set((char*)0)`
 * Fixed segfault after `variant.set(serialized((char*)0))`
+* Detect `IncompleteInput` in `false`, `true`, and `null`
 
 v6.8.0-beta (2019-01-30)
 -----------